artifactory_du/requirements.txt
requests
dohq-artifactory
hurry.filesize